in_time_engine migration
Included the original main.cpp and Test.h from the original project. Resolved some more issues relating to conditional compilation. Project continue to compile successfully. It runs without crashing but it the game screen shows a black screen. It needs to be fixed. The editor window on the other hand renders and works correctly.