dependency build automation
Now there is only one bat file that MSVC interacts with to build dependencies named build_dependencies.bat. This script call a script per dependency. It needs to receive a list of macros from MSVC to conditionally run the other scripts. Right now, this script works and is able to compile all dependencies prior to linking.