README
README is too long. I reviwed it to correct some errors and replace hard written paths with hyperlinks. README will be restructured in next commit. BUILD.md and more will be added and move sections of the README to this and some other files to shorten the file and not overwhelm first time readers.