Input System and Backend Implementattion
GLFW Engine shoud not be a thing. It does not scale well with the goal of making the engine cross platform. I migrated the input system of GLFW Engine to In Time Engine. It now has a standardized way to forward input to this system. The function is named feedAction for the Input and JoystickManager class. GLFW Engine will be reduced to a backend implementation that forwards the GLFW data to the engine. This way, no matter what input system you use, you can access your input the same way everywhere. All you have to do is select the backend you want to use.